Core Help - Alerts Topic Updates
HelpStartConceptsAlerts updated to include mention and link to passive
scanning as a source of alerts, as well as mentioning that scripts can
be a source of alerts.
HelpUiTabsAlerts removed redundant paragraph about alert sources, as
it's already covered in the concepts page. (Also cleaned up line breaks
vs paragraphs.)

Fixes zaproxy/zaproxy#2243