Update Break Point dialog page
Rename the page (title) to Add/Edit Break Point dialog, it allows to add
and edit the break point, also describe the fields that compose the HTTP
break point.
Update TOC and index to reflect the rename of the page.

Related to zaproxy/zaproxy#4477 - Enable help in Add/Edit Break Point
dialogue