Merge help of request editor dialogues
Merge the help of Resend dialogue into the page of Manual Request Editor
dialogue, both editors have the same behaviour/functionality, the only
difference is that one starts with a default request while the other
with an existing one.
Update labels of the menu item that allows to open the Manual Request
Editor dialogue with an existing request, part of zaproxy/zaproxy#3818.
Remove the Resend help page and its references (or replace them with the
Manual Request Editor page, as appropriate).